Refactor out an interpolated string
Refactor out an interpolated string being passed to a call to `debug`. This is to that when requiring client-side packages such as `@feathersjs/rest-client` and `@feathersjs/authentication-client` and loading the bundled code in IE 11 we don't get any invalid token errors (i.e. interpolated strings are not supported by IE 11).